refactor: user comments resource
